+++ packages/terra-form/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update Field documentation
+* Change NumberField defaultValue propType to accept string or numeric values
 
 1.1.0 - (July 18, 2017)
 ------------------